c# test cases
Here are the test cases for the c# code. They use NUnit.
To add NUnit to the vs community:
right click on Solution -> manage nuget packages for solution
Select Browse at the top and search for NUnit. 
Select the packages shown in the screen shot and install. (If the latest version is not working correctly try installing the previous version)